Fix Room return arguments for insert

+++ app/src/main/java/com/sarmaru/mihai/jetpackfundamentals/db/DogBreedDao.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Dao
 public interface DogBreedDao {
     @Insert
-    List<Integer> insertAll(DogBreed... dogs);
+    List<Long> insertAll(DogBreed... dogs);
 
     @Query("SELECT * FROM dogbreed")
     List<DogBreed> getAllDogs();
